garden-sentinel/pyproject.toml
[project]
name = "garden-sentinel"
version = "0.1.0"
description = "AI-powered garden security system to protect chickens from predators"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Garden Sentinel Team" }
]
keywords = ["security", "camera", "ai", "detection", "raspberry-pi", "yolo"]

dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
    "paho-mqtt>=1.6.1",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
edge = [
    "picamera2>=0.3.12",
    "gpiozero>=2.0",
    "pigpio>=1.78",
]
coral = [
    "pycoral>=2.0.0",
    "tflite-runtime>=2.14.0",
]
training = [
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "albumentations>=1.3.0",
    "matplotlib>=3.7.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
garden-sentinel-server = "garden_sentinel.server.main:main"
garden-sentinel-edge = "garden_sentinel.edge.main:main"
garden-sentinel-train = "garden_sentinel.training.train:main"
garden-sentinel-collect = "garden_sentinel.training.data_collector:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["garden_sentinel"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.uv.workspace]
members = ["packages/*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
